Bug fixes and code cleanup

+++ include/grph.h
@@ -8,7 +8,7 @@ typedef struct OverlapGraph {
     char ***edges;
 } overlap_graph_t;
 
-bool has_overlap(const char *seq1, const char *seq2, unsigned int k);
-overlap_graph_t *calc_overlap_graph(fasta_t *fasta_seqs, unsigned int k);
+bool has_overlap(const char *seq1, const char *seq2, size_t k);
+overlap_graph_t *calc_overlap_graph(fasta_t *fasta_seqs, size_t k);
 
 void grph(const char *filename);
